requirements.txt
Flask
mysql-connector-python
PyJWT
bcrypt